+++ apps/ungdomsytelse-deltaker/src/api/deltakelse-perioder/getDeltakelsePerioder.ts
@@ -1,6 +1,7 @@
 import { handleApiError } from '@navikt/ung-common';
 import { Deltakelse } from '@navikt/ung-deltakelse-opplyser-api-deltaker';
 
+import { logApiErrorFaro } from '../../apps/innsyn/utils/apiErrorLogger';
 import { DeltakelsePeriode, deltakelsePerioderSchema } from '../../types/DeltakelsePeriode';
 
 export const getDeltakelsePerioder = async (): Promise<DeltakelsePeriode[]> => {
@@ -8,6 +9,8 @@ export const getDeltakelsePerioder = async (): Promise<DeltakelsePeriode[]> => {
         const { data } = await Deltakelse.hentAlleMineDeltakelser();
         return deltakelsePerioderSchema.parse(data);
     } catch (e) {
-        throw handleApiError(e, 'getAlleMineDeltakelser');
+        const parsedApiError = handleApiError(e, 'getDeltakelsePerioder');
+        logApiErrorFaro('getDeltakelsePerioder', parsedApiError);
+        throw parsedApiError;
     }
 };