+++ apps/ungdomsytelse-deltaker/src/components/deltaker-info-loader/DeltakerInfoLoader.tsx
@@ -51,7 +51,6 @@ const DeltakerInfoLoader = () => {
         const søkerError = getErrorInfoToLog(søker.error);
         const deltakelsePerioderError = getErrorInfoToLog(deltakelsePerioder.error);
         logApiError(ApiErrorKey.oppstartsinfo, { søkerError, deltakelsePerioderError });
-        logFaroError('DeltakerInfoLoader.Error', JSON.stringify({ søkerError, deltakelsePerioderError }));
         return <HentDeltakerErrorPage error="Feil ved henting av info" />;
     }